Update de-DE.json
Miscellaneous fixes, mainly smoother syntax, consistency changes and different wording. I also inverted the order of the thread priority dropdown strings at the end, now "0" is "highest" because I was assuming that that's how it's supposed to be, if it's wrong that part can just be left like it was.