Filling all the gaps in uk-UA.json (again)
I've decided to use "н/в" as a translation to "N/A", even though it can be interpreted as a short version of the word "невизначений", which means "undefined". Also, I guess "play-layer" means the function that lets you to play the level right in the editor (I might be wrong, my apologies in that case), but it could be a bit difficult to express with only a couple of words in Ukrainian, so after thinking about it for some time, I've chosen to use the word for a level ("рівень") instead. If there are any questions — DM me in Discord (@whoisamiaq)